% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constants}
\alias{constants}
\alias{on_god}
\alias{cap}
\title{Vibin Constants}
\format{
An object of class \code{logical} of length 1.

An object of class \code{logical} of length 1.
}
\usage{
on_god

cap
}
\description{
Facts
}
\keyword{datasets}
