% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rent_free.R
\name{rent_free}
\alias{rent_free}
\title{Rent Free}
\usage{
rent_free(x, y)
}
\arguments{
\item{x, y}{Vectors to test}
}
\value{
elements of x in y
}
\description{
get elements of x living rent-free in y
}
\examples{
{
    a <- 1:2
    b <- 1:3
    rent_free(a, b)
}
}
