% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibing.R
\name{vibing}
\alias{vibing}
\title{Vibing}
\usage{
vibing(.x, .f, ..., .progress = FALSE)
}
\arguments{
\item{.x}{a list or atomic vector}

\item{.f}{function to apply}

\item{...}{arguments passed to .f}

\item{.progress}{show progress bar?}
}
\value{
a vector the same length as .x
}
\description{
just keep vibing. Apply a function to each element of a vector
}
\examples{
{
    z <- list(x = 1:5, y = 6:10)
    vibing(z, mean)
}
}
